Revert "fix: rename from solutionformat -> solution"
This reverts commit 7771a6b09071f0061b9e9dd0b08ae4926de8d0e3.
