+++ integration/oasis/solutionformat/doc.go
@@ -1,7 +1,7 @@
 /*
-package solution defines internal foramt for recording single solution result.
+package solutionformat defines internal foramt for recording single solution result.
 It isolates internal calculation logic and final restful response:
 - Internal algorithm will calculate result in this format.
 - In the end, oasis will use this format to generate restful response.
 */
-package solution
+package solutionformat
